Move Keysym/Unicode mapping table to separate file
Make it easier to see what's code and what's just a lot of data.
